llm-observability/worker/requirements.txt
APScheduler==3.10.4
psycopg2-binary==2.9.9
httpx==0.27.2
pydantic-settings==2.5.2
python-json-logger==2.0.7
